Move our main code to a lib
- In `main.rs`, all we do is call `execute()` on our library
