+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["signet-mob <signet.mob@gmail.com>"]
 edition = "2018"
-name = "serial-port-reader"
+name = "serial-port-reader-writer"
 version = "0.2.1"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -16,3 +16,7 @@ toml = "0.5"
 [dependencies.serde]
 features = ["derive"]
 version = "1.0"
+
+[dev-dependencies]
+pretty_assertions = "0.7"
+rstest = "0.7"
